Add  for micro-sd argument, add .gitignore file, check directories before creation, add camera config.txt keyvalues

+++ get_raspbian.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-MICROSDDEVICE="" # /dev/sd?
+MICROSDDEVICE=$1 # /dev/sd?
 RASPIANSHA256="a50237c2f718bd8d806b96df5b9d2174ce8b789eda1f03434ed2213bbca6c6ff" # NEEDS
 # Exit script if a non-zero exit code
 set -e
@@ -10,7 +10,9 @@ uname -a | grep Linux || ( echo "not on a linux box, lets stop here" && exit -1 
 
 # Get raspbian 
 
-curl -L -o "raspbian.zip" --url "https://downloads.raspberrypi.org/raspbian_lite_latest"
+if [[ ! -f raspbian.zip ]]; then
+    curl -L -o "raspbian.zip" --url "https://downloads.raspberrypi.org/raspbian_lite_latest";
+fi
 
 sha256sum raspbian.zip | grep ${RASPIANSHA256} || \
  ( echo "ShaSum not correct...." && exit )
